+++ quilt-sync/src-tauri/src/main.rs
@@ -129,7 +129,6 @@ fn main() {
             commands::package_uninstall,
             commands::reset_local,
             commands::reveal_in_file_browser,
-            commands::set_origin,
             commands::set_remote,
             commands::package_create,
             commands::setup,